Move organization tests to separate file
